Update aaa.c
Removed un-needed typecast as we are casting to the same type!

+++ src/aaa.c
@@ -4,5 +4,5 @@
 #include <stdio.h>
 
 uint8_t *aaa(uint8_t *buffer) {
-  return write8(buffer, 0x37, (unsigned long)sizeof(*buffer) / sizeof(uint8_t));
+  return write8(buffer, 0x37, sizeof(*buffer) / sizeof(uint8_t));
 }
